[prmr#528] Redesign CanvasFont as a Singleton.
