+++ src/org/jetuml/rendering/StringRenderer.java
@@ -27,17 +27,11 @@ import java.util.Map;
 
 import org.jetuml.annotations.Flyweight;
 import org.jetuml.annotations.Immutable;
-import org.jetuml.application.UserPreferences;
-import org.jetuml.application.UserPreferences.IntegerPreference;
-import org.jetuml.application.UserPreferences.IntegerPreferenceChangeHandler;
 import org.jetuml.geom.Dimension;
 import org.jetuml.geom.Rectangle;
 
 import javafx.geometry.VPos;
 import javafx.scene.canvas.GraphicsContext;
-import javafx.scene.text.Font;
-import javafx.scene.text.FontPosture;
-import javafx.scene.text.FontWeight;
 import javafx.scene.text.TextAlignment;
 
 /**
@@ -50,7 +44,7 @@ import javafx.scene.text.TextAlignment;
 @Flyweight
 public final class StringRenderer
 {
-	private static final CanvasFont CANVAS_FONT = new CanvasFont();
+	private static final CanvasFont CANVAS_FONT = CanvasFont.instance();
 	
 	private static final Dimension EMPTY = new Dimension(0, 0);
 	private static final int DEFAULT_HORIZONTAL_TEXT_PADDING = 7;
@@ -208,32 +202,6 @@ public final class StringRenderer
         formattedString.deleteCharAt(formattedString.length() - 1);
     	return formattedString.toString();
 	}
-
-	private TextAlignment getTextAlignment()
-	{		
-		if ( aAlign.isLeft() )
-		{
-			return TextAlignment.LEFT;
-		}
-		else if ( aAlign.isHorizontallyCentered() )
-		{
-			return TextAlignment.CENTER;
-		}
-		return TextAlignment.RIGHT;
-	}
-	
-	private VPos getTextBaseline()
-	{
-		if ( aAlign.isBottom() )
-		{
-			return VPos.BASELINE;
-		}
-		else if ( aAlign.isTop() )
-		{
-			return VPos.TOP;
-		}
-		return VPos.CENTER;
-	}
 	
 	/**
      * Draws the string inside a given rectangle.
@@ -266,18 +234,7 @@ public final class StringRenderer
 		}
 		
 		pGraphics.translate(pRectangle.getX(), pRectangle.getY());
-		if( aBold && aItalic )
-		{
-			RenderingUtils.drawText(pGraphics, textX, textY, pString.trim(), CANVAS_FONT.getBoldItalic());
-		}
-		else if( aItalic )
-		{
-			RenderingUtils.drawText(pGraphics, textX, textY, pString.trim(), CANVAS_FONT.getItalic());
-		}
-		else
-		{
-			CANVAS_FONT.drawString(pGraphics, textX, textY, pString.trim(), aBold);
-		}
+		RenderingUtils.drawText(pGraphics, textX, textY, pString.trim(), CANVAS_FONT.getFont(aBold, aItalic));
 		
 		if(aUnderlined && pString.trim().length() > 0)
 		{
@@ -287,7 +244,7 @@ public final class StringRenderer
 			if( aAlign.isHorizontallyCentered() )
 			{
 				xOffset = dimension.width()/2;
-				yOffset = CANVAS_FONT.fontSize()/2 + 1;
+				yOffset = dimension.height()/2;
 			}
 			else if( aAlign.isRight() )
 			{
@@ -305,129 +262,29 @@ public final class StringRenderer
 		pGraphics.setTextAlign(oldAlign);
 	}
 	
-	/**
-	 * Responsible for performing more rudimentary operations involving font,
-	 * as well as being synchronized with the user's current font.
-	 */
-	private static final class CanvasFont implements IntegerPreferenceChangeHandler
+	private TextAlignment getTextAlignment()
+	{		
+		if ( aAlign.isLeft() )
+		{
+			return TextAlignment.LEFT;
+		}
+		else if ( aAlign.isHorizontallyCentered() )
+		{
+			return TextAlignment.CENTER;
+		}
+		return TextAlignment.RIGHT;
+	}
+	
+	private VPos getTextBaseline()
 	{
-
-		private Font aFont;
-		private Font aFontBold;
-		private Font aFontItalic;
-		private Font aFontBoldItalic;
-		private FontMetrics aFontMetrics;
-		private FontMetrics aFontBoldMetrics;
-		private FontMetrics aFontItalicMetrics;
-		private FontMetrics aFontBoldItalicMetrics;
-
-		private CanvasFont()
+		if ( aAlign.isBottom() )
 		{
-			refreshAttributes();
-			UserPreferences.instance().addIntegerPreferenceChangeHandler(this);
+			return VPos.BASELINE;
 		}
-
-		private Font getFont(boolean pBold)
+		else if ( aAlign.isTop() )
 		{
-			if ( pBold )
-			{
-				return aFontBold;
-			}
-			return aFont;
+			return VPos.TOP;
 		}
-		
-		public Font getItalic()
-		{
-			return aFontItalic;
-		}
-		
-		public Font getBoldItalic()
-		{
-			return aFontBoldItalic;
-		}
-
-		private FontMetrics getFontMetrics(boolean pBold, boolean pItalic)
-		{
-			if( pBold && pItalic )
-			{
-				return aFontBoldItalicMetrics;
-			}
-			else if( pBold )
-			{
-				return aFontBoldMetrics;
-			}
-			else if( pItalic )
-			{
-				return aFontItalicMetrics;
-			}
-			return aFontMetrics;
-		}
-
-		/**
-		 * Returns the dimension of a given string.
-		 * @param pString The string to which the bounds pertain.
-		 * @return The dimension of the string
-		 */
-		public Dimension getDimension(String pString, boolean pBold, boolean pItalic)
-		{
-			return getFontMetrics(pBold, pItalic).getDimension(pString);
-		}
-		
-		/**
-		 * Returns the height of a string including the leading space.
-		 * 
-		 * @param pString The string.
-		 * @param pBold Whether the text is in bold.
-		 * @return The height of the string.
-		 */
-		public int getHeight(String pString, boolean pBold, boolean pItalic)
-		{
-			return getFontMetrics(pBold, pItalic).getHeight(pString);
-		}
-
-		/**
-		 * Draws the string on the graphics context at the specified position.
-		 * @param pGraphics The graphics context
-		 * @param pTextX The x-position of the string
-		 * @param pTextY The y-position of the string
-		 * @param pString The canvas on which to draw the string
-		 * @param pBold If the text should be bold
-		 */
-		public void drawString(GraphicsContext pGraphics, int pTextX, int pTextY, String pString, boolean pBold)
-		{
-			RenderingUtils.drawText(pGraphics, pTextX, pTextY, pString, getFont(pBold));
-		}
-
-		/**
-		 * Returns the font size the user currently specifies.
-		 * @return The font size
-		 */
-		public int fontSize()
-		{
-			return (int) Math.round(aFont.getSize());
-		}
-
-		@Override
-		public void integerPreferenceChanged(IntegerPreference pPreference) 
-		{
-			if ( pPreference == IntegerPreference.fontSize && aFont.getSize() != UserPreferences.instance().getInteger(pPreference) )
-			{
-				refreshAttributes();
-			}
-
-		}
-
-		private void refreshAttributes()
-		{
-			aFont = Font.font("System", UserPreferences.instance().getInteger(IntegerPreference.fontSize));
-			aFontBold = Font.font(aFont.getFamily(), FontWeight.BOLD, aFont.getSize());
-			aFontItalic = Font.font(aFont.getFamily(), FontPosture.ITALIC, aFont.getSize());
-			aFontBoldItalic = Font.font(aFont.getFamily(), FontWeight.BOLD, FontPosture.ITALIC, aFont.getSize());
-			aFontMetrics = new FontMetrics(aFont);
-			aFontBoldMetrics = new FontMetrics(aFontBold);
-			aFontItalicMetrics = new FontMetrics(aFontItalic);
-			aFontBoldItalicMetrics = new FontMetrics(aFontBoldItalic);
-		}
-
+		return VPos.CENTER;
 	}
 }
